requirements.txt
discord.py==2.4.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
apscheduler==3.10.4
python-dotenv==1.0.1
aiosqlite==0.20.0